% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{check_clean_userspace}
\alias{check_clean_userspace}
\title{Deprecated: use \code{\link[=audit_userspace]{audit_userspace()}} instead}
\usage{
check_clean_userspace(pkg = ".", check_output = tempfile("dircheck"))
}
\arguments{
\item{pkg}{Path to the package to audit.}

\item{check_output}{Where to store check outputs (defaults to a tempfile).}
}
\value{
See \code{\link[=audit_userspace]{audit_userspace()}}.
}
\description{
Deprecated: use \code{\link[=audit_userspace]{audit_userspace()}} instead
}
\keyword{internal}
