% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkhelper-package.R
\docType{package}
\name{checkhelper-package}
\alias{checkhelper}
\alias{checkhelper-package}
\title{checkhelper: Deal with Check Outputs}
\description{
Deal with packages 'check' outputs and reduce the risk of rejection by 'CRAN' by following policies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thinkr-open.github.io/checkhelper/}
  \item \url{https://github.com/ThinkR-open/checkhelper}
  \item Report bugs at \url{https://github.com/ThinkR-open/checkhelper/issues}
}

}
\author{
\strong{Maintainer}: Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Authors:
\itemize{
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})
  \item Arthur Bréant \email{arthur@thinkr.fr} (\href{https://orcid.org/0000-0003-1668-0963}{ORCID})
  \item Murielle Delmotte \email{murielle@thinkr.fr} (\href{https://orcid.org/0000-0002-1339-2424}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
