% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_values.R
\name{find_missing_values}
\alias{find_missing_values}
\title{Find missing 'return' tag when function exported}
\usage{
find_missing_values(
  package.dir = ".",
  roclets = NULL,
  load_code = NULL,
  clean = FALSE
)
}
\arguments{
\item{package.dir}{Location of package top level directory. Default is
working directory.}

\item{roclets}{Character vector of roclet names to use with package.
The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}.}

\item{load_code}{A function used to load all the R code in the package
directory. The default, \code{NULL}, uses the strategy defined by
the \code{load} roxygen option, which defaults to \code{\link[roxygen2:load]{load_pkgload()}}.
See \link[roxygen2]{load} for more details.}

\item{clean}{If \code{TRUE}, roxygen will delete all files previously
created by roxygen before running each roclet.}
}
\value{
Tibble of all functions with information whether documentation is correct or not
}
\description{
Find missing 'return' tag when function exported
}
\examples{
\dontrun{
find_missing_values()
}
}
