% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{find_nonascii_files}
\alias{find_nonascii_files}
\title{Deprecated: use \code{\link[=audit_ascii]{audit_ascii()}} instead}
\usage{
find_nonascii_files(
  path = ".",
  scope = c("R", "tests", "vignettes", "man", "DESCRIPTION", "NAMESPACE"),
  ignore_ext = c("png", "jpg", "jpeg", "gif", "rds", "rda", "rdata", "pdf", "ico", "svg"),
  size_limit = 5e+05
)
}
\arguments{
\item{scope}{Character vector of subdirectories / files to scan.}

\item{ignore_ext}{Extensions to skip (binary assets, snapshots).}

\item{size_limit}{Skip files larger than this many bytes.}
}
\value{
See \code{\link[=audit_ascii]{audit_ascii()}}.
}
\description{
Deprecated: use \code{\link[=audit_ascii]{audit_ascii()}} instead
}
\keyword{internal}
