% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_data_info}
\alias{get_data_info}
\title{Deprecated: internal helper used by \code{\link[=fix_dataset_doc]{fix_dataset_doc()}}}
\usage{
get_data_info(name, description, source)
}
\arguments{
\item{name}{Name of the file that exists in \verb{data/}.}

\item{description}{Description for the data.}

\item{source}{Source of data.}
}
\value{
A list of information from a data.frame.
}
\description{
Deprecated: internal helper used by \code{\link[=fix_dataset_doc]{fix_dataset_doc()}}
}
\keyword{internal}
