% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_no_visible}
\alias{get_no_visible}
\title{Deprecated: use \code{\link[=audit_globals]{audit_globals()}} instead}
\usage{
get_no_visible(path = ".", checks, ...)
}
\arguments{
\item{path}{Path to package.}

\item{checks}{Output of \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} if already computed.}

\item{...}{Other parameters passed to \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}.}
}
\value{
See \code{\link[=audit_globals]{audit_globals()}}.
}
\description{
Deprecated: use \code{\link[=audit_globals]{audit_globals()}} instead
}
\keyword{internal}
