% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_notes}
\alias{get_notes}
\title{Deprecated: internal helper used by \code{\link[=audit_globals]{audit_globals()}}}
\usage{
get_notes(path = ".", checks, ...)
}
\arguments{
\item{path}{Path to package.}

\item{checks}{Output of \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}} if already computed.}

\item{...}{Other parameters passed to \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}.}
}
\value{
A tibble of notes from \verb{R CMD check} containing global-related
information.
}
\description{
Deprecated: internal helper used by \code{\link[=audit_globals]{audit_globals()}}
}
\keyword{internal}
