% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{print_globals}
\alias{print_globals}
\title{Deprecated: use \code{\link[=fix_globals]{fix_globals()}} instead}
\usage{
print_globals(globals, path = ".", ..., message = TRUE)
}
\arguments{
\item{globals}{A list as issued from \code{\link[=get_no_visible]{get_no_visible()}} or empty.}

\item{path}{Path to package.}

\item{...}{Other parameters passed to \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck()}}.}

\item{message}{Logical. Whether to return message with content (Default)
or return as list.}
}
\value{
See \code{\link[=fix_globals]{fix_globals()}}.
}
\description{
Deprecated: use \code{\link[=fix_globals]{fix_globals()}} instead
}
\keyword{internal}
