% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_as_cran.R
\name{check_as_cran}
\alias{check_as_cran}
\title{Check your package with CRAN default values for environment variables}
\usage{
check_as_cran(
  pkg = ".",
  check_dir,
  scratch,
  Ncpus = 1,
  as_command = FALSE,
  clean_before = TRUE,
  open = FALSE
)
}
\arguments{
\item{pkg}{pkg directory to check}

\item{check_dir}{Where to store check outputs}

\item{scratch}{Where to store temporary files (cleaned after)}

\item{Ncpus}{Number of Cpus}

\item{as_command}{Whether to run the check as Linux command line, instead of directly in R}

\item{clean_before}{Whether to delete the previous check_dir}

\item{open}{Whether to open the check dir at the end of the process}
}
\value{
An object containing errors, warnings, and notes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
check_dir <- tempfile("example")
# Check the current directory
check_as_cran(check_dir = check_dir)
# Open directory with all outputs
utils::browseURL(check_dir)
}
}
