% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clean_userspace.R
\name{check_clean_userspace}
\alias{check_clean_userspace}
\title{check_clean_userspace}
\usage{
check_clean_userspace(pkg, check_dir = tempfile("dircheck"))
}
\arguments{
\item{pkg}{Path to package to check}

\item{check_dir}{Path to directory where to store check results}
}
\value{
data.frame of files that are left after checks
}
\description{
check_clean_userspace
}
\examples{
\dontrun{
check_clean_userspace()
}
\dontrun{
check_clean_userspace()
}
}
