% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clean_userspace.R
\name{check_clean_userspace}
\alias{check_clean_userspace}
\title{check_clean_userspace}
\usage{
check_clean_userspace(pkg = ".", check_output = tempfile("dircheck"))
}
\arguments{
\item{pkg}{Path to package to check}

\item{check_output}{Path to directory where to store check results}
}
\value{
data.frame of files that are left after checks
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# This runs a check of the current package
all_files <- check_clean_userspace()
all_files
}
}
