% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example_pkg.R
\name{create_example_pkg}
\alias{create_example_pkg}
\title{Create a package example producing notes and errors}
\usage{
create_example_pkg(
  path = tempfile(pattern = "pkg-"),
  with_functions = TRUE,
  with_extra_notes = FALSE
)
}
\arguments{
\item{path}{Path where to store the example package}

\item{with_functions}{Logical. Whether there will be functions or not (with notes)}

\item{with_extra_notes}{Logical. Whether there are extra notes or not}
}
\value{
Path where the example package is stored.
}
\description{
Create a package example producing notes and errors
}
\examples{
create_example_pkg()
}
