% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_info.R
\name{get_data_info}
\alias{get_data_info}
\title{Get infos form .Rdata inside data folder}
\usage{
get_data_info(name, description, source)
}
\arguments{
\item{name}{name of file}

\item{description}{description for the data}

\item{source}{source of data}
}
\value{
list of infos form data
}
\description{
Get infos form .Rdata inside data folder
}
\examples{
path_data <- tempfile(pattern = "data-")
dir.create(path_data)
path_rda <- file.path(path_data, "iris.rda")
save(iris, file = path_rda)
withr::with_dir(
  path_data, {
   get_data_info("iris", "Iris data frame", source = "ThinkR")
  })

# Clean userspace
unlink(path_data, recursive = TRUE)

}
