% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data_doc.R
\name{use_data_doc}
\alias{use_data_doc}
\title{Documentation of .rda in package}
\usage{
use_data_doc(
  name,
  prefix = "doc_",
  description = "Description",
  source = "Source"
)
}
\arguments{
\item{name}{Name of your data. only the name,}

\item{prefix}{Add prefix for the name of R script}

\item{description}{Add a description}

\item{source}{Add a source}
}
\description{
Documentation of .rda in package
}
\examples{
\dontrun{
use_data_doc("my_data", description = "Desc of my_data", source = "Here my source")
}
}
